Add option to destroy damageable on death
First I thought to use an event for the OnDie Action, but there would be
no warranty that the destroy command will be the last in the chain of
actions happening OnDie. After this, I decided to add a boolean that
indicates if the damageable should be destroyed after death. This
happening after the call to OnDie.Invoke()