Change wave start logic
Now the wave starts after all the enemies are death.

When an enemy dies it fires an event in the channel EnemyDiedEvent

The EnemySpawnerController listens the channel EnemyDiedEvent, and
when the channel is fired it decreases the number of enemies alive in
the register. If the amount of enemies alive are 0 or less it starts a
new wave of enemies